refactor: Extract message, scale modules
