+++ js/main.js
@@ -1,9 +1,9 @@
+import { setFileFieldChange } from './upload-form.js';
 import { getData } from './api.js';
 import { renderThumbnails } from './thumbnail-gallery.js';
 import { setThumbnailFilters } from './filter.js';
-import { showDownloadError } from './util.js';
 import { setFullSizeModalHandlers } from './full-size-modal.js';
-import { setFileFieldChange } from './upload-form.js';
+import { showDownloadError } from './message.js';
 
 const initPictureGallery = async () => {
   try {
